Initial commit
This starts off a spike to connect Go to the Gazebo emulator; connecting
to an emulator seems like a natural first thing in building a robotics
library.

I started spiking in the direction of using Gazebos client library,
linking in the shared library directly. However, while I was able to get
that to work for gazebo 6, gazebo 7 has changed their build system to
cmake, and I'm not cool enough to understand how to translate the cmake
dependency definitions into something cgo can understand.

Looking at the c++ client code, it's an extremely simple protocol, and
the existing implementation is not written in a way that makes me think
linking to it from any language would be a performance benefit.

Hence, rather than dig into the hell hole of connecting this large
dependency chain via cgo, I'm spiking out a pure go client; this should
be a useful way to learn to program go as well.