Swapped off of the Gazebo train, onto the Morse train
I think the diff in this commit is rather self-explanatory. Getting
going with Gazebo's protocol took two days of reverse-engineering and
lead to hundreds of files, hundreds of lines of code to do anything;

Meanwhile, Morse took.. 18 lines of code, less than 5 minutes to
understand what to do. While I don't know much about either simulator, I
know a good amount about about software, and this to me is an excellent
indication of the thought that's gone into the Morse simulator.